Revert "Item Drag & Drop"
