+++ apps/assassin/src/components/song/AddSongDialog.tsx
@@ -54,12 +54,13 @@ function FormField({ label, htmlFor, required, icon, children }: FormFieldProps)
 }
 
 export function AddSongDialog({ seasonId, open, onOpenChange, onSubmit }: AddSongDialogProps) {
-  const { validateSongData } = useSongLogic(seasonId);
+  const { validateSongData, getNextSortOrder } = useSongLogic(seasonId);
 
   const { formState, isProcessing, handleSubmit, handleChangeField } = useSongHandlers({
     mode: "create",
     initialData: {
       seasonId,
+      sortOrder: getNextSortOrder(),
     },
   });
