+++ apps/assassin/src/features/song/hooks/useSongLogic.ts
@@ -1,13 +1,10 @@
-import { useCallback, useMemo } from "react";
+import { useCallback } from "react";
 import type { Song } from "../schema";
 import type { SongFormData } from "./useSongForm";
 import { useSongsBySeason } from "../queries/useSongsBySeason";
-import { createSongSortOrderHelpers } from "../utils/songSortOrderHelpers";
 
 export const useSongLogic = (seasonId: string, initialSongs: Song[] = []) => {
   const { data: songs = initialSongs } = useSongsBySeason(seasonId, initialSongs);
-
-  const { sortedSongs } = useMemo(() => createSongSortOrderHelpers(songs), [songs]);
 
   const isNameExists = useCallback(
     (name: string, excludeId?: string): boolean => {
@@ -16,6 +13,45 @@ export const useSongLogic = (seasonId: string, initialSongs: Song[] = []) => {
       );
     },
     [songs],
+  );
+
+  const getNextSortOrder = useCallback((): number => {
+    if (songs.length === 0) return 1;
+
+    const maxSortOrder = Math.max(...songs.map((song) => Number(song.sortOrder)));
+    return maxSortOrder + 1;
+  }, [songs]);
+
+  const getSortOrderBetween = useCallback(
+    (beforeId: string | null, afterId: string | null): number => {
+      const sortedSongs = [...songs].sort((a, b) => Number(a.sortOrder) - Number(b.sortOrder));
+
+      if (!beforeId && !afterId) {
+        return getNextSortOrder();
+      }
+
+      if (!beforeId) {
+        // Insert at beginning
+        const firstSong = sortedSongs[0];
+        return firstSong ? Number(firstSong.sortOrder) - 1 : 1;
+      }
+
+      if (!afterId) {
+        // Insert at end
+        return getNextSortOrder();
+      }
+
+      // Insert between two songs
+      const beforeSong = songs.find((s) => s.id === beforeId);
+      const afterSong = songs.find((s) => s.id === afterId);
+
+      if (!beforeSong || !afterSong) {
+        return getNextSortOrder();
+      }
+
+      return Math.floor((Number(beforeSong.sortOrder) + Number(afterSong.sortOrder)) / 2);
+    },
+    [songs, getNextSortOrder],
   );
 
   const validateSongData = useCallback(
@@ -80,8 +116,12 @@ export const useSongLogic = (seasonId: string, initialSongs: Song[] = []) => {
     [songs],
   );
 
+  const sortedSongs = [...songs].sort((a, b) => Number(a.sortOrder) - Number(b.sortOrder));
+
   return {
     isNameExists,
+    getNextSortOrder,
+    getSortOrderBetween,
     validateSongData,
     reorderSongs,
